+++ Test/Migration/Services/MigrationProgressServiceTest.php
@@ -11,7 +11,7 @@ use Doctrine\DBAL\Connection;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
-use Shopware\Core\Framework\DataAbstractionLayer\Indexing\MessageQueue\IndexerMessageSender;
+use Shopware\Core\Framework\DataAbstractionLayer\Indexing\EntityIndexerRegistry;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Write\EntityWriter;
@@ -214,7 +214,7 @@ class MigrationProgressServiceTest extends TestCase
                 $this->mediaFileRepo,
                 $this->salesChannelRepo,
                 $this->themeRepo,
-                $this->getContainer()->get(IndexerMessageSender::class),
+                $this->getContainer()->get(EntityIndexerRegistry::class),
                 $this->getContainer()->get(ThemeService::class),
                 $this->getContainer()->get(MappingService::class),
                 $this->getContainer()->get('cache.object'),